proxyless_gaze/training/face_detection/requirements.txt
# TODO: Update with exact module version
numpy
torch>=1.7
opencv_python
loguru
scikit-image
tqdm
torchvision
Pillow
thop
ninja
tabulate
tensorboard

# verified versions
onnx
onnxruntime
onnx-simplifier